refactor(routes): extract session-list cache helpers from api/routes.py (#1907)
